refactor(nms): Move projects out of fbc_js_core
